// src/types.ts
export interface iPhone {
  id: string;
  name: string;
  year: string;
  releaseDate: string;
  processor: string;
  display: string;
  camera: string;
  ram: string;
  imageUrl: string;
  dimensions?: string;
  storage: string[];
  colors?: string[];
  features: string[];
  price: string;
  description: string;
} 